Version 1: As received from Brainy007
This Kodi Addon is deiscussed here:

https://forum.kodi.tv/showthread.php?tid=31170

and this the last version that Brainy007 shared. It works pretty well! But it needs some work in some areas. Notably missing a key mix setting and alos a really annoying problem in that when the mix is added to Kodi, it is rendered only as file paths in the web interface and the on the phone app. Not until you view the playlist in Kodi itself, does Kodi finally scan those files for metadata and do all the vews on the playlist update with songe titles and artist names and even cover images on the phone app.